New OpenCLGridOperators extensions class
- fixed Performance by using singleton only
- new extension for extending the OpenCLGridOperators class for new,
own OpenCL kernel functions, etc.
- improved documentation
